Split ulox test runner tests into own fixture
